web-app-serve/floyd_requirements.txt
Pillow
scipy
numpy
flask
dicom
Pillow
